+++ Projects/Flow/Sources/App/AppFlow.swift
@@ -55,9 +55,9 @@ public class AppFlow: Flow {
     }
 
     private func presentLoginView() -> FlowContributors {
-        let loginFlow = LoginFlow(container: self.container)
+        let authFlow = AuthFlow(container: self.container)
 
-        Flows.use(loginFlow, when: .created) { [weak self] root in
+        Flows.use(authFlow, when: .created) { [weak self] root in
             UIView.transition(
                 with: self!.window,
                 duration: 0.5,
@@ -69,7 +69,7 @@ public class AppFlow: Flow {
 
         return .one(
             flowContributor: .contribute(
-                withNextPresentable: loginFlow,
+                withNextPresentable: authFlow,
                 withNextStepper: OneStepper(
                     withSingleStep: PiCKStep.loginIsRequired
                 )
